Started work on MqMessageStream.
